+++ database/config_test.go
@@ -218,213 +218,3 @@ func TestIndexStore(t *testing.T) {
 		require.Len(t, list, len(idxcfgs)-1)
 	})
 }
-
-func TestFieldConstraintsInfer(t *testing.T) {
-	tests := []struct {
-		name      string
-		got, want FieldConstraints
-		fails     bool
-	}{
-		{
-			"No change",
-			[]*FieldConstraint{{Path: document.NewPath("a"), Type: document.IntegerValue}},
-			[]*FieldConstraint{{Path: document.NewPath("a"), Type: document.IntegerValue}},
-			false,
-		},
-		{
-			"Array",
-			[]*FieldConstraint{{Path: document.NewPath("a", "0"), Type: document.IntegerValue}},
-			[]*FieldConstraint{
-				{Path: document.NewPath("a"), Type: document.ArrayValue, IsInferred: true, InferredBy: []document.Path{document.NewPath("a", "0")}},
-				{Path: document.NewPath("a", "0"), Type: document.IntegerValue},
-			},
-			false,
-		},
-		{
-			"Document",
-			[]*FieldConstraint{{Path: document.NewPath("a", "b"), Type: document.IntegerValue}},
-			[]*FieldConstraint{
-				{Path: document.NewPath("a"), Type: document.DocumentValue, IsInferred: true, InferredBy: []document.Path{document.NewPath("a", "b")}},
-				{Path: document.NewPath("a", "b"), Type: document.IntegerValue},
-			},
-			false,
-		},
-		{
-			"Complex path",
-			[]*FieldConstraint{{Path: document.NewPath("a", "b", "3", "1", "c"), Type: document.IntegerValue}},
-			[]*FieldConstraint{
-				{Path: document.NewPath("a"), Type: document.DocumentValue, IsInferred: true, InferredBy: []document.Path{document.NewPath("a", "b", "3", "1", "c")}},
-				{Path: document.NewPath("a", "b"), Type: document.ArrayValue, IsInferred: true, InferredBy: []document.Path{document.NewPath("a", "b", "3", "1", "c")}},
-				{Path: document.NewPath("a", "b", "3"), Type: document.ArrayValue, IsInferred: true, InferredBy: []document.Path{document.NewPath("a", "b", "3", "1", "c")}},
-				{Path: document.NewPath("a", "b", "3", "1"), Type: document.DocumentValue, IsInferred: true, InferredBy: []document.Path{document.NewPath("a", "b", "3", "1", "c")}},
-				{Path: document.NewPath("a", "b", "3", "1", "c"), Type: document.IntegerValue},
-			},
-			false,
-		},
-		{
-			"Overlaping constraints",
-			[]*FieldConstraint{
-				{Path: document.NewPath("a", "b"), Type: document.IntegerValue},
-				{Path: document.NewPath("a", "c"), Type: document.IntegerValue},
-			},
-			[]*FieldConstraint{
-				{Path: document.NewPath("a"), Type: document.DocumentValue, IsInferred: true, InferredBy: []document.Path{document.NewPath("a", "b"), document.NewPath("a", "c")}},
-				{Path: document.NewPath("a", "b"), Type: document.IntegerValue},
-				{Path: document.NewPath("a", "c"), Type: document.IntegerValue},
-			},
-			false,
-		},
-		{
-			"Same path inferred and non inferred: inferred first",
-			[]*FieldConstraint{
-				{Path: document.NewPath("a", "b"), Type: document.IntegerValue},
-				{Path: document.NewPath("a"), Type: document.DocumentValue},
-			},
-			[]*FieldConstraint{
-				{Path: document.NewPath("a"), Type: document.DocumentValue},
-				{Path: document.NewPath("a", "b"), Type: document.IntegerValue},
-			},
-			false,
-		},
-		{
-			"Same path inferred and non inferred: inferred last",
-			[]*FieldConstraint{
-				{Path: document.NewPath("a"), Type: document.DocumentValue},
-				{Path: document.NewPath("a", "b"), Type: document.IntegerValue},
-			},
-			[]*FieldConstraint{
-				{Path: document.NewPath("a"), Type: document.DocumentValue},
-				{Path: document.NewPath("a", "b"), Type: document.IntegerValue},
-			},
-			false,
-		},
-		{
-			"Complex case",
-			[]*FieldConstraint{
-				{Path: document.NewPath("a", "b", "3", "1", "c"), Type: document.DocumentValue},
-				{Path: document.NewPath("a", "b", "3", "1", "c", "d"), Type: document.IntegerValue},
-				{Path: document.NewPath("a", "b", "2"), Type: document.IntegerValue},
-			},
-			[]*FieldConstraint{
-				{Path: document.NewPath("a"), Type: document.DocumentValue, IsInferred: true,
-					InferredBy: []document.Path{document.NewPath("a", "b", "3", "1", "c"), document.NewPath("a", "b", "3", "1", "c", "d"), document.NewPath("a", "b", "2")}},
-				{Path: document.NewPath("a", "b"), Type: document.ArrayValue, IsInferred: true,
-					InferredBy: []document.Path{document.NewPath("a", "b", "3", "1", "c"), document.NewPath("a", "b", "3", "1", "c", "d"), document.NewPath("a", "b", "2")}},
-				{Path: document.NewPath("a", "b", "3"), Type: document.ArrayValue, IsInferred: true,
-					InferredBy: []document.Path{document.NewPath("a", "b", "3", "1", "c"), document.NewPath("a", "b", "3", "1", "c", "d")}},
-				{Path: document.NewPath("a", "b", "3", "1"), Type: document.DocumentValue, IsInferred: true,
-					InferredBy: []document.Path{document.NewPath("a", "b", "3", "1", "c"), document.NewPath("a", "b", "3", "1", "c", "d")}},
-				{Path: document.NewPath("a", "b", "3", "1", "c"), Type: document.DocumentValue},
-				{Path: document.NewPath("a", "b", "3", "1", "c", "d"), Type: document.IntegerValue},
-				{Path: document.NewPath("a", "b", "2"), Type: document.IntegerValue},
-			},
-			false,
-		},
-		{
-			"Same path, different constraint",
-			[]*FieldConstraint{
-				{Path: document.NewPath("a", "b"), Type: document.IntegerValue},
-				{Path: document.NewPath("a", "b"), Type: document.DoubleValue},
-			},
-			nil,
-			true,
-		},
-		{
-			"Inferred constraint first, conflict with non inferred",
-			[]*FieldConstraint{
-				{Path: document.NewPath("a", "b"), Type: document.IntegerValue},
-				{Path: document.NewPath("a"), Type: document.IntegerValue},
-			},
-			nil,
-			true,
-		},
-		{
-			"Non inferred constraint first, conflict with inferred",
-			[]*FieldConstraint{
-				{Path: document.NewPath("a"), Type: document.IntegerValue},
-				{Path: document.NewPath("a", "b"), Type: document.IntegerValue},
-			},
-			nil,
-			true,
-		},
-	}
-
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			got, err := test.got.Infer()
-			if test.fails {
-				require.Error(t, err)
-			} else {
-				require.NoError(t, err)
-				require.Equal(t, test.want, got)
-			}
-		})
-	}
-}
-
-func TestFieldConstraintsAdd(t *testing.T) {
-	tests := []struct {
-		name  string
-		got   FieldConstraints
-		add   FieldConstraint
-		want  FieldConstraints
-		fails bool
-	}{
-		{
-			"Same path",
-			[]*FieldConstraint{{Path: document.NewPath("a"), Type: document.IntegerValue}},
-			FieldConstraint{Path: document.NewPath("a"), Type: document.IntegerValue},
-			nil,
-			true,
-		},
-		{
-			"Duplicate primary key",
-			[]*FieldConstraint{{Path: document.NewPath("a"), IsPrimaryKey: true, Type: document.IntegerValue}},
-			FieldConstraint{Path: document.NewPath("b"), IsPrimaryKey: true, Type: document.IntegerValue},
-			nil,
-			true,
-		},
-		{
-			"Different path",
-			[]*FieldConstraint{{Path: document.NewPath("a"), Type: document.IntegerValue}},
-			FieldConstraint{Path: document.NewPath("b"), Type: document.IntegerValue},
-			[]*FieldConstraint{
-				{Path: document.NewPath("a"), Type: document.IntegerValue},
-				{Path: document.NewPath("b"), Type: document.IntegerValue},
-			},
-			false,
-		},
-		{
-			"Default value conversion, typed constraint",
-			[]*FieldConstraint{{Path: document.NewPath("a"), Type: document.IntegerValue}},
-			FieldConstraint{Path: document.NewPath("b"), Type: document.IntegerValue, DefaultValue: document.NewDoubleValue(5)},
-			[]*FieldConstraint{
-				{Path: document.NewPath("a"), Type: document.IntegerValue},
-				{Path: document.NewPath("b"), Type: document.IntegerValue, DefaultValue: document.NewIntegerValue(5)},
-			},
-			false,
-		},
-		{
-			"Default value conversion, untyped constraint",
-			[]*FieldConstraint{{Path: document.NewPath("a"), Type: document.IntegerValue}},
-			FieldConstraint{Path: document.NewPath("b"), DefaultValue: document.NewIntegerValue(5)},
-			[]*FieldConstraint{
-				{Path: document.NewPath("a"), Type: document.IntegerValue},
-				{Path: document.NewPath("b"), DefaultValue: document.NewDoubleValue(5)},
-			},
-			false,
-		},
-	}
-
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			err := test.got.Add(&test.add)
-			if test.fails {
-				require.Error(t, err)
-			} else {
-				require.NoError(t, err)
-				require.Equal(t, test.want, test.got)
-			}
-		})
-	}
-}
